scripts/data_cataloger/pip-requirements.txt
wheel==0.37.1
click==8.0.3
PyYAML==6.0
pytest==7.0.0
pytest-xdist==2.5.0
pytest-mock==3.7.0
pytest-cov==3.0.0
coverage==6.3.1
Faker==13.11.1
pytz==2022.1
